[projectfirma/#1409] Track Funding Source Amount
* The Funding Source Amount input uses currency formatting

+++ Database/ReleaseScript/0190 - Add FundingSourceAmount.sql
@@ -1,2 +1,2 @@
 ALTER TABLE [dbo].[FundingSource]
-ADD FundingSourceAmount FLOAT NULL;
+ADD FundingSourceAmount MONEY NULL;
